requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
Django==3.1.2
django-oauth-toolkit==1.3.3
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.11
